requirements.txt
# requirements.txt
beautifulsoup4>=4.12.2
requests>=2.31.0
pyyaml>=6.0
openai>=1.0.0
tidalapi>=0.8.3
# Optional packages for local model processing
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.20.0
bitsandbytes>=0.39.0